test: split world and oauth Postgres suites
